+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### Bug Fixes
 * **terra-text-input** changed color of text input with readonly attribute
+* **terra-dynamic-form** added tooltip for category picker
+* **terra-category-picker** added input for tooltip
 
 <a name="2.2.2"></a>
 # 2.2.2 (25.04.2018)